LANG_002a Fronted overview connected to the backed  [SM,LA]

+++ src/main/java/at/tugraz/sw20asd/lang/ui/VocabularyAccess.java
@@ -9,7 +9,7 @@ public interface VocabularyAccess {
 
     Integer addVocabulary(Vocabulary vocabulary);
 
-    Vocabulary getVocabulary(int id);
+    Vocabulary getVocabulary(long id);
 
     Collection<Vocabulary> getAllVocabularies();
 }
